fix: 운동 참여 컨트롤러 분리
